Fixed A bug
Fixed a bug in function WOETable(), where if one of the levels in a factor variable is not actually in the data, possibly because of splitting the data into training and test, then, creating the good bad count for that variable will throw an error. This is now, passed through a try() statement, to suppress the error.